Move "Attach Debugger" command to a more appropriate group, and make it a package command rather than an editor command, so that it does not require REPL to be in focus to invoke. Also display and focus the REPL window after successfully attaching.
Add workspace-related commands to the REPL toolbar (except for "Restart R", because REPL already has a restart button).
